+++ index.html
@@ -9,6 +9,7 @@
 
 <body>
   <div id="app"></div>
-  <script type="module" src="script.js"></script>
+  <!-- <script type="module" src="main.js"></script> -->
+  <script src="./dist/main.js"></script>
 </body>
 </html>
