+++ registration.js
@@ -63,13 +63,13 @@ export const renderReg = ({ fetchPromise }) => {
           login: loginInputElement.value,
           password: passwordInputElement.value
         })
-                  .then((responseData) => {
+            .then((responseData) => {
             console.log(responseData);
             setToken(responseData.user.token);
             setLoggedInUserName(responseData.user.name);
           })
           .then(() => {
-            fetchPromise();
+            fetchPromise()
           })
           .catch((error) => {
             console.error(error);
