+++ renderComment.js
@@ -1,5 +1,6 @@
-import { appComment } from "./script.js";
-import { formatDate } from "./data.js";
+import { appComment } from "./main.js";
+// import { formatDate } from "./data.js";
+import { format } from "date-fns";
 
 export function renderComment({ comments }) {
   const appElement = document.getElementById("app");
@@ -11,11 +12,15 @@ export function renderComment({ comments }) {
   const commentsHtml = comments.map((comment, index) => {
     const commentTextQuotes = comment.comment.replaceAll("QUOTE_BEGIN", "<div class='quote'>").replaceAll("QUOTE_END", "</div>");
     const commentNameSafe = comment.name.replace(/</g, '&lt;').replace(/>/g, '&gt;');
+    const now = new Date();
+    format(now, 'yyyy-MM-dd hh.mm.ss');
+    const createDate = format(now, 'yyyy-MM-dd hh.mm.ss');
+
     return `
       <li class="comment" id="comment-add" data-index=${index}>
         <div class="comment-header">
           <div>${commentNameSafe}</div>
-          <div>${comment.data}</div>
+          <div>${createDate}</div>
         </div>
         <div class="comment-body">
           <div class="comment-text" >${commentTextQuotes}</div>
@@ -95,6 +100,8 @@ export function renderComment({ comments }) {
   nameInputElement.classList.add("inactive");
 
   buttonElement.addEventListener("click", () => {
+    const now = new Date();
+    format(now, 'yyyy-MM-dd hh.mm.ss');
 
     nameInputElement.classList.remove("error")
     if (nameInputElement.value === "") {
@@ -110,7 +117,7 @@ export function renderComment({ comments }) {
 
     const userName = nameInputElement.value;
     const userComment = textInputElement.value;
-    const userData = formatDate(new Date());
+    const userData = format(now, 'yyyy-MM-dd hh.mm.ss');;
 
     appComment(userName, userComment, userData);
   });
